update the modified code for slippy map, the current code is based on the master branch

+++ Assets/EsriPrototypeLab/Scripts/Extensions.cs
@@ -19,7 +19,7 @@ using UnityEngine;
 namespace Esri.PrototypeLab.HoloLens.Unity {
     public static class Extensions {
         public static Tile ToTile(this Coordinate coordinate, int zoom) {
-            var latrad = coordinate.Latitude * Mathf.Deg2Rad;
+            var latrad = (float)coordinate.Latitude * Mathf.Deg2Rad;
             var n = Mathf.Pow(2, zoom);
             var x = (coordinate.Longitude + 180) / 360 * n;
             var y = (1f - Mathf.Log(Mathf.Tan(latrad) + (1 / Mathf.Cos(latrad))) / Mathf.PI) / 2.0 * n;
